+++ database/migrations/2025_01_07_033120_create_user_profiles_table.php
@@ -1,0 +1,36 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::create('user_profiles', function (Blueprint $table) {
+            $table->uuid('profile_id')->primary();
+            $table->uuid('user_id');
+            $table->string('last_name')->nullable(false);
+            $table->string('first_name')->nullable(false);
+            $table->string('middle_name')->nullable();
+            $table->string('gender')->nullable(false);
+            $table->timestamps();
+            
+            // Foreign Key Assignment
+            $table->foreign('user_id')->references('user_id')->on('custom_user_tables')->onDelete('restrict');
+        });
+
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::dropIfExists('user_profiles');
+    }
+};